logger: initialize buf to an empty string to avoid atoi() reading garbage
In the very unlikely case that buffer buf is not populated with data
because a read is not performed, we end up with several atoll or atoi
calls on a garbage buffer.  Fix this by setting the 0th element of buf
with '\0' to avoid parsing garbage.

Detected by CoverityScan, CID#208902 ("Unitialized scalar variable")

Signed-off-by: Colin Ian King <colin.king@canonical.com>