perf_msr: re-work read_msr to return error code as an int
Currently read_msr returns the msr value as a uint64_t and this
is being compared to -1 to see if it is an error. This potentially
can provide a false error indication if the msr being read returns
the uint64_t value 0xffffffffffffffff. Fix this by returning an
int status and by passing the value read as a 3rd arguent to the
function.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>